changed datatype from to integer

+++ calculator.py
@@ -1,6 +1,6 @@
 # Ask the user to input two numbers and an operation
-num1 = float(input("Enter the first number: "))
-num2 = float(input("Enter the second number: "))
+num1 = int(input("Enter the first number: "))
+num2 = int(input("Enter the second number: "))
 operation = input("Enter operation (+, -, *, /): ")
 
 # Perform the chosen operation and display the result
